fix(via): provide the option to not alloc on exit
