+++ examples/cookies/src/main.rs
@@ -118,5 +118,5 @@ async fn main() -> Result<ExitCode, Error> {
     // Add a route that responds with a greeting message.
     app.at("/hello/:name").respond(via::get(hello));
 
-    via::start(app).listen(("127.0.0.1", 8080)).await
+    Ok(via::start(app).listen(("127.0.0.1", 8080)).await?)
 }
