+++ examples/hello-world/src/main.rs
@@ -25,5 +25,5 @@ async fn main() -> Result<ExitCode, Error> {
     // Define a route that listens on /hello/:name.
     app.at("/hello/:name").respond(via::get(hello));
 
-    via::start(app).listen(("127.0.0.1", 8080)).await
+    Ok(via::start(app).listen(("127.0.0.1", 8080)).await?)
 }
