+++ examples/shared-state/src/main.rs
@@ -93,5 +93,5 @@ async fn main() -> Result<ExitCode, Error> {
     // Add the `totals` responder to the endpoint GET /totals.
     app.at("/totals").respond(via::get(totals));
 
-    via::start(app).listen(("127.0.0.1", 8080)).await
+    Ok(via::start(app).listen(("127.0.0.1", 8080)).await?)
 }
