--- src/server/error.rs
@@ -1,32 +1,0 @@
-use std::error::Error;
-use std::fmt::{self, Display, Formatter};
-use std::io;
-
-#[derive(Debug)]
-pub enum ServerError {
-    Io(io::Error),
-    Hyper(hyper::Error),
-}
-
-impl Error for ServerError {}
-
-impl Display for ServerError {
-    fn fmt(&self, f: &mut Formatter) -> fmt::Result {
-        match self {
-            Self::Io(e) => Display::fmt(e, f),
-            Self::Hyper(e) => Display::fmt(e, f),
-        }
-    }
-}
-
-impl From<io::Error> for ServerError {
-    fn from(e: io::Error) -> Self {
-        Self::Io(e)
-    }
-}
-
-impl From<hyper::Error> for ServerError {
-    fn from(e: hyper::Error) -> Self {
-        Self::Hyper(e)
-    }
-}
